Write pseudo-code for game()
Wrote pseudo-code to establish a rough idea of how to implement game().
Going with a net round win approach because of noTieLogic()'s return
type already being a 1 / 0 for a win, and just needing to adjust that.
I wrote a point to keep track of round is considered valid because I'm
planning to use prompt() to keep track of user input, and it's possible
for the user to hit OK after putting no input, or CANCEL regardless of
input. I'm thinking of having hit OK because counted towards a round,
and CANCEL not.