Write declareMultipleRound() logic
Since I'm taking in netOutcome as the input parameter, which is to be of
a number type. I'm going to expect as per pseudo-code to be greater than
zero, zero, or less than zero.

- Greater than zero will indicate that the player has net victory
- Equal to zero will indicate the player and computer reached net tie.
  This is a rare possibility in a game of 5 rounds, but is possible.
- Less than zero will indicate that the player has net lost

The equals to zero case is also a necessity since I'm wanting to give
more flexibility to game() by letting it take in any amount of rounds.

The last case of default in the switch logic is just to cover the case
where the argument passed into the declareMultipleRound()'s parameter
is not the expected value.