Write roundNotCanceled() logic
It's pretty simple, but following on the heels of the logic done for
declareMultipleRound(), I think having each round hold a value relative
to the player's netOutcome is smart - thus the expected roundOutcome
values individually being +1 / 0 / -1, as per original game() comments.

I'm deciding to go with an arbitrary value of -2 from roundOutcome to
effectively be a flag to detect if the user canceled the round or not.
If the round was canceled, I'll have roundOutcome be assigned to -2.
Otherwise, I'll have it be assigned to +1 / 0 / -1.