Rename getCom...Choice() to getComputerSelection()
Renamed it to such to keep a consistent naming scheme between 'choice'
and 'selection'. The difference will be that for getting player and
computer selections, will be more of a "rough input". Then for the
player and computer choices, that'll be the vocabulary used for when
their selections are reformatted for playRound() logic.