Write getPlayerSelection() logic
I also removed the related TODO comment within game(), and modified the
1st argument passed into playRound() in game() to call the new logic
written in getPlayerSelection().

I'm checking for null because that's what's output by prompt() when the
user cancels. Something to note is that when the user doesn't put an
input value into the prompt, then hits okay, it returns a empty string.

As stated before, I'll be treating the "null" values as a skip, so it
won't decrement the rounds counter in game(). Then empty string values
will be treated as a free tie round, since the user hit OK to start the
round. This is an arbitrary design decision, to be clear.

The return statement is written using the ternary operator '?' since
the original if else statment was really verbose for how short it was.
I personally still can understand the one line ternary code fine, and
in exchange, the overall amount of code lines in logic.js is minimized.