Add playRound() pseudo-code to itself
Having realized that playRound() was really difficult to skim to figure
out what was going on, I want to improve its readability to a similar
level of game() so that I understand what's going on.

This is because playRound() is inherently not a function that does only
a single thing, because to play a round requires multiple sub-tasks to
be done. That being said, I'd like to allocate its sub-tasks according
to the TODO's drafted in the comments, being:

 1. Improve readability of playRound() by abstraction
 2. Add a dedicated isTie() method due to needing to consider "" input
 3. Abstract declaration logic to a method like declareSingleRound()
 4. Change output parameter to be suitable for roundOutcome in game()

The #2 TODO needs to consider "" input, which is once again possible
whenever the user clicks OK on the prompt without typing anything in.
Refer to previous commit, commit #8955f05 for the context about why.