Write isTie() and modify related logic
Again, as per precedence, writing in ternary operator to preserve
overall readability of whole logic.js file, while keeping the desired
readability of isTie(), relative to its if else format originally.

The logic can be described as:
 1. If playerChoice is "" aka prompt input empty after OK, return true
 2. If prompt input is not empty after OK, check for normal tie

I don't have to worry about the input arguments being some invalid
value because both get...Selection() methods handle this already.

Modified the logic of variable result in playRound() because if there
is a tie, I want to treat that as round as a value 0 relative to the
netOutcome of potentially multiple rounds.