Update noTieLogic() logic
Added a default case to noTieLogic to account for the cases where
the player input into prompt is something like null, which is converted
into "null" - or something purely random like "abc123" and needs to be
caught.

These cases are treated as invalid rounds, as per roundNotCanceled().
Thus, not being counted towards the rounds counter in game().