Write declareSingleRound() and update playRound()
Originally, I just moved all of the original declaration logic from
playRound() to declareSingleRound(). Then I remembered that I was
treating the case values differently due to the higher-level process
described in game() to handle multiple rounds. Thus, the code has been
adjusted accordingly within declareSingleRound(), with a default case
thrown in there just incase something truly unexpected happens.

The extra parameters playerChoice and computerChoice are there so that
I still meet the spirit of the original TOP requirements for the method
playRound() - but declareSingleRound() and declareMultipleRound() are
inevitably very similar otherwise.

The notable difference though aside from parameters, is that I chose
to output single round outcomes to the console, rather than alert. This
is because from a game perspective, I only want the user to be alerted
to the overall outcome at the end of the rounds, but still want the
individual round outcomes outputted to communicate results during the
game itself for myself.

Also marked the TODO comments within playRound() to DONE accordingly.