Rename netOutcome, netRoundWins to netRoundOutcome
These two variables were supposed to originally be the same value, but
I mixed up my vocabulary for it as I implemented the rest of the code.
I decided to best preserve readability between game() and the given
variable's relationships to the other methods and variables (e.g.
roundOutcome) to rename it to a combination of both names.