Fix logic errors in noTieLogic() and playRound()
In noTieLogic(), I updated the outcomes for the false choices of the
ternary operator cases from 0 to -1, since noTieLogic() is only supposed
to handle wins and losses. The 0 was a relic from the original logic
used in playRound() before I even considered creating game(), and the
functions that build up to it.

In playRound(), I forgot to pass in arguments into isTie(). I learned
today that if you don't pass in the arguments, that the default values
will be undefined, hence undefined === undefined passing as true, and
giving the output of "rock ties paper", because I still initialized
playerChoice and computerChoice properly in playRound() earlier.