Fix declareSingleRound() & declareMultipleRound()
What happened was that for the last case of each function before the
default case, I did not put a break statement in it like I should have.
This resulted in the console.error() from the default case being output
whenever I also passes the last case condition above it, for each
function respectively - this is a property of the switch statement, and
why it's important to include break statements in every case.

The only exception is for the last case at the bottom,which does not
typically require a break. Unless there's a default case, in which the
last non-default case would need it, like this situation.