Write 3rd & 4th tests for Cypress Test Suite
Technically, 4x tests since it's 3, 3A, 4, 4A, but I digress.
The tests 3 & 4 are just to check that I can initialize the two
respective gamemodes I plan to have for this rps overhaul. The tests 3A
and 4A are to verify that I can play through each respective gamemode
with no issues with the end outcome, even if the computer's choices are
randomized.