Write 5th (batch) of tests for Cypress Test Suite
This batch of 3x tests - 5, 5A, 5B - along with the 1x helper method,
are here to test the intended features and mechanics of a reset button
for the rock, paper, scissors page once overhauled.

In general, most data-cy attributes used more than once within the it()
portions of the test spec file are being converted to variable format to
reduce repetitive code, and make it easier to modify the code as needed.