Write out pseudocode for RPS logic flows
Wrote out pseudocode for the overall RPS logic flow plan, and the
details to go with each specific RPS logic flow for the play button,
reset button, and theme button, which are going to be my main trigger
points for various logics.

I've come to the realization that even though how I wrote my original
code was perfectly fine for what was required of me, now that I have a
chance to rework and expand upon this project, I'll be going beyond the
spec requirements, as I've thought over this logic flow to make it my
own. I'm going to essentially rebuild the html and js files, and add
a css file while I'm at it.

The goal is to have every single test I've written up in the Cypress
test suite to pass without issues or flakiness. Trying out test driven
development so that I can have some level of assurance that I won't take
this project way out of the scope I have originally intended.