Establish pseudo-structure for styles.css
Setting up comments for where I want what types of CSS to go.
I'm hoping to practice categorical CSS, as alphabetical feels a little
chaotic for me to keep track of for what properties go with what from
a logical perspective. As I'm working on this project overhaul alone,
I don't have to worry about maintaining a consistent team-wide pool of
knowledge, so structuring the CSS how it makes sense to me, in a overall
consistent manner, is of higher utility to me, over a universal way
of interpreting CSS.