Allow the user to forfeit the game early
Before, the user was not allowed to forfeit the game early, even if they
clicked cancel or pressed Esc. This is wrong because it was intended for
users to be able to leave early in such a case.

Now, the user is allowed to forfeit the game early, and a warning log
is displayed indicating their action, and the subsequent result.